requirements.txt
python-docx~=1.1
Jinja2~=3.1
PyYAML~=6.0
pypdf~=3.11
reportlab~=4.0 